+++ pkg/lib/helper.go
@@ -1,14 +1,9 @@
-package commands
+package lib
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"io/fs"
-	"os"
-	"path/filepath"
-	"regexp"
-	"strings"
 
 	"github.com/slok/sloth/internal/app/generate"
 	"github.com/slok/sloth/internal/log"
@@ -20,15 +15,13 @@ import (
 	pluginenginesli "github.com/slok/sloth/internal/pluginengine/sli"
 	pluginengineslo "github.com/slok/sloth/internal/pluginengine/slo"
 	storagefs "github.com/slok/sloth/internal/storage/fs"
-	"github.com/slok/sloth/pkg/common/model"
 )
 
-func createPluginLoader(ctx context.Context, logger log.Logger, paths []string) (*storagefs.FilePluginRepo, error) {
-	fss := []fs.FS{
-		plugin.EmbeddedDefaultSLOPlugins,
-	}
-	for _, p := range paths {
-		fss = append(fss, os.DirFS(p))
+func createPluginLoader(ctx context.Context, logger log.Logger, pluginsFS []fs.FS) (*storagefs.FilePluginRepo, error) {
+	// We should load at least the Sloth embedded default ones.
+	fss := append([]fs.FS{}, pluginsFS...)
+	if len(fss) == 0 {
+		fss = append(fss, plugin.EmbeddedDefaultSLOPlugins)
 	}
 
 	pluginsRepo, err := storagefs.NewFilePluginRepo(logger, pluginenginesli.PluginLoader, pluginengineslo.PluginLoader, fss...)
@@ -37,73 +30,6 @@ func createPluginLoader(ctx context.Context, logger log.Logger, paths []string) 
 	}
 
 	return pluginsRepo, nil
-}
-
-func discoverSLOManifests(logger log.Logger, exclude, include *regexp.Regexp, path string) ([]string, error) {
-	logger = logger.WithValues(log.Kv{"svc": "SLODiscovery"})
-
-	paths := []string{}
-	err := filepath.Walk(path, func(path string, info fs.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-
-		if info.IsDir() {
-			return nil
-		}
-
-		// Directories and non YAML files don't need to be handled.
-		extension := strings.ToLower(filepath.Ext(path))
-		if info.IsDir() || (extension != ".yml" && extension != ".yaml") {
-			return nil
-		}
-
-		// Filter by exclude or include (exclude has preference).
-		if exclude != nil && exclude.MatchString(path) {
-			logger.Debugf("Excluding path due to exclude filter %s", path)
-			return nil
-		}
-		if include != nil && !include.MatchString(path) {
-			logger.Debugf("Excluding path due to include filter %s", path)
-			return nil
-		}
-
-		// If we reach here, path discovered.
-		paths = append(paths, path)
-
-		return nil
-	})
-
-	if err != nil {
-		return nil, fmt.Errorf("could not find files recursively: %w", err)
-	}
-
-	return paths, nil
-}
-
-func mapCmdPluginToModel(ctx context.Context, jsonPlugins []string) ([]model.PromSLOPluginMetadata, error) {
-	type jsonPlugin struct {
-		ID       string          `json:"id"`
-		Config   json.RawMessage `json:"config"`
-		Priority int             `json:"priority"`
-	}
-
-	plugins := []model.PromSLOPluginMetadata{}
-	for _, jp := range jsonPlugins {
-		p := &jsonPlugin{}
-		err := json.Unmarshal([]byte(jp), p)
-		if err != nil {
-			return nil, fmt.Errorf("could not load cmd plugin json config %q: %w", jp, err)
-		}
-
-		plugins = append(plugins, model.PromSLOPluginMetadata{
-			ID:       p.ID,
-			Config:   p.Config,
-			Priority: p.Priority,
-		})
-	}
-
-	return plugins, nil
 }
 
 func createDefaultSLOPlugins(logger log.Logger, disableRecordings, disableAlerts bool) ([]generate.SLOProcessor, error) {